Renamed DependencyNode and DependencyTree
Removing the ' from the file name, causing issues with symbolsource.
Made error.css a embedded resource and error-test.htm is no longer content which was then pulled into nuget package
